# frontend/openapi/openapi.json
{
  "components": {
    "schemas": {
      "PreparacionOut": {
        "additionalProperties": false,
        "description": "Readiness: todas las dependencias están disponibles.",
        "examples": [
          {
            "comprobaciones": {
              "base_datos": "ok",
              "migraciones": "ok"
            },
            "estado": "listo"
          }
        ],
        "properties": {
          "comprobaciones": {
            "additionalProperties": {
              "const": "ok",
              "type": "string"
            },
            "title": "Comprobaciones",
            "type": "object"
          },
          "estado": {
            "const": "listo",
            "title": "Estado",
            "type": "string"
          }
        },
        "required": [
          "estado",
          "comprobaciones"
        ],
        "title": "PreparacionOut",
        "type": "object"
      },
      "Problema": {
        "additionalProperties": true,
        "description": "Detalle de problema; admite extensiones como `errores` o `comprobaciones`.",
        "examples": [
          {
            "detail": "El recurso solicitado no existe.",
            "instance": "urn:uuid:3f2b8c1e-6d4a-4f7e-9a51-2b0c8d7e6f10",
            "status": 404,
            "title": "Recurso no encontrado",
            "type": "no_encontrado"
          }
        ],
        "properties": {
          "detail": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Detail"
          },
          "instance": {
            "description": "`urn:uuid:<correlation_id>` de la petición.",
            "title": "Instance",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "integer"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "type": {
            "description": "Código de error de docs/04-api.md §12.",
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "type",
          "title",
          "status",
          "instance"
        ],
        "title": "Problema",
        "type": "object"
      },
      "SaludOut": {
        "additionalProperties": false,
        "description": "Liveness: el proceso responde.",
        "examples": [
          {
            "estado": "vivo"
          }
        ],
        "properties": {
          "estado": {
            "const": "vivo",
            "title": "Estado",
            "type": "string"
          }
        },
        "required": [
          "estado"
        ],
        "title": "SaludOut",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "Agente Correo API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/v1/salud": {
      "get": {
        "description": "Indica que el proceso responde; no consulta dependencias.",
        "operationId": "vivo_api_v1_salud_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SaludOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problema"
                }
              }
            },
            "description": "Petición inválida (`validacion`)."
          },
          "404": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problema"
                }
              }
            },
            "description": "Recurso no encontrado (`no_encontrado`)."
          },
          "500": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problema"
                }
              }
            },
            "description": "Error interno (`error_interno`)."
          }
        },
        "summary": "Liveness",
        "tags": [
          "salud"
        ]
      }
    },
    "/api/v1/salud/listo": {
      "get": {
        "description": "Comprueba las dependencias; 503 `no_listo` si alguna falla.",
        "operationId": "listo_api_v1_salud_listo_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PreparacionOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problema"
                }
              }
            },
            "description": "Petición inválida (`validacion`)."
          },
          "404": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problema"
                }
              }
            },
            "description": "Recurso no encontrado (`no_encontrado`)."
          },
          "500": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problema"
                }
              }
            },
            "description": "Error interno (`error_interno`)."
          },
          "503": {
            "content": {
              "application/problem+json": {
                "example": {
                  "comprobaciones": {
                    "base_datos": "ok",
                    "migraciones": "falla"
                  },
                  "detail": "Una o más dependencias no están disponibles.",
                  "instance": "urn:uuid:3f2b8c1e-6d4a-4f7e-9a51-2b0c8d7e6f10",
                  "status": 503,
                  "title": "Servicio no listo",
                  "type": "no_listo"
                },
                "schema": {
                  "$ref": "#/components/schemas/Problema"
                }
              }
            },
            "description": "Alguna dependencia no está disponible (`no_listo`)."
          }
        },
        "summary": "Readiness",
        "tags": [
          "salud"
        ]
      }
    }
  }
}